pdf-link-checker: Fix thread launch logic
The threading logic was completely broken. We waited 1-sec if there
was too many threads, but after the sleep we launched another one,
thus exceeding the limit.

So we always launch the maximum thread count, assigning one URL to each,
using a protected URL list. Each thread dies when this list is empty.

Signed-off-by: Ezequiel Garcia <ezequiel.garcia@free-electrons.com>